+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.6.1] - 2022-05-19
 ###  Added
 - Prevent duplicate middleware call.
 - Added method (`Route::middleware`) to call middleware individualy.
@@ -15,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.6.0] - 2022-05-15
 ### Added
-- Add `RouteDispatcher::class` dispatch request route.
+- Added `RouteDispatcher::class` dispatch request route.
 
 ### Changed
 - `Router::class` run throw `RouteDispatcher::class`.